[jsk_perception] Add advanced morphological transformations
