+++ jsk_perception/CMakeLists.txt
@@ -114,6 +114,11 @@ jsk_perception_nodelet(src/add_mask_image.cpp "jsk_perception/AddMaskImage" "add
 jsk_perception_nodelet(src/multiply_mask_image.cpp "jsk_perception/MultiplyMaskImage" "multiply_mask_image")
 jsk_perception_nodelet(src/dilate_erode_mask_image.cpp "jsk_perception/DilateMaskImage" "dilate_mask_image")
 jsk_perception_nodelet(src/dilate_erode_mask_image.cpp "jsk_perception/ErodeMaskImage" "erode_mask_image")
+jsk_perception_nodelet(src/morphological_operator.cpp "jsk_perception/Opening" "opening")
+jsk_perception_nodelet(src/morphological_operator.cpp "jsk_perception/Closing" "closing")
+jsk_perception_nodelet(src/morphological_operator.cpp "jsk_perception/MorphologicalGradient" "morphological_gradient")
+jsk_perception_nodelet(src/morphological_operator.cpp "jsk_perception/TopHat" "top_hat")
+jsk_perception_nodelet(src/morphological_operator.cpp "jsk_perception/BlackHat" "black_hat")
 jsk_perception_nodelet(src/rect_to_mask_image.cpp "jsk_perception/RectToMaskImage" "rect_to_mask_image")
 jsk_perception_nodelet(src/mask_image_to_roi.cpp "jsk_perception/MaskImageToROI" "mask_image_to_roi")
 jsk_perception_nodelet(src/mask_image_to_rect.cpp "jsk_perception/MaskImageToRect" "mask_image_to_rect")
